Refactor comment handling: move _Comment and _CommentBanner to utils; update imports for consistency
