--- CityOfBinds/src/game/bind_file/comments/__init__.py
@@ -1,4 +1,0 @@
-from .comment import _Comment
-from .comment_banner import _CommentBanner
-
-__all__ = ["_Comment", "_CommentBanner"]
