+++ CityOfBinds/src/game/utils/__init__.py
@@ -1,0 +1,7 @@
+from .comments.comment import _Comment
+from .comments.comment_banner import _CommentBanner
+
+__all__ = [
+    "_Comment",
+    "_CommentBanner",
+]
